mail_send/requirements.txt
amqp==5.1.1
asgiref==3.6.0
async-timeout==4.0.3
billiard==4.1.0
celery==5.3.4
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
colorlog==6.7.0
cron-descriptor==1.4.0
distlib==0.3.6
Django==4.2
django-celery-beat==2.5.0
django-celery-results==2.5.1
django-log-request-id==2.1.0
django-redis==5.4.0
django-timezone-field==6.0.1
djangorestframework==3.14.0
filelock==3.11.0
kombu==5.3.2
platformdirs==3.2.0
prompt-toolkit==3.0.39
python-crontab==3.0.0
python-dateutil==2.8.2
pytz==2023.3.post1
redis==4.6.0
six==1.16.0
sqlparse==0.4.3
tzdata==2023.3
vine==5.0.0
virtualenv==20.21.0
wcwidth==0.2.8
